requirements.txt
asgiref==3.5.2
Django==4.1.3
django-countries==7.4.2
django-phonenumber-field==7.0.0
phonenumbers==8.13.0
Pillow==9.3.0
sqlparse==0.4.3
typing_extensions==4.4.0
